refactor JobDataModel to JobId
